requirements.txt
praw==7.3.0
pandas==1.2.5
sklearn==0.24.2
hydra-core==1.1.0
hydra-optuna-sweeper==1.1.0
pytorch-lightning==1.3.7
mlflow==1.18.0
wandb==0.10.32
fastapi==0.65.2
uvicorn==0.14.0